+++ Easy/C/Project Teams.c
@@ -1,10 +1,8 @@
 #include <stdio.h>
 
-//Compiler version gcc  6.3.0
-
 int main() {
-  int N;
-  scanf("%d", &N);
-  printf("%d", N/3);
-  return 0;
+	unsigned long N;
+	scanf("%lu", &N);
+	printf("%lu", N / 3);
+	return 0;
 }
